requirements.txt
streamlit
langchain
langchain-groq
langchain-core
sentence-transformers
faiss-cpu
duckduckgo-search
